% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTimeClus.R
\docType{data}
\name{airparif}
\alias{airparif}
\title{Real spatio-temporal data: airparif}
\description{
airparif$obs describes 101 days in 2015 by indicating the quantity of NO2 at 9 sites around Paris during 24 hours.
}
\details{
airparif$map indicates the locations of the 9 sites  around Paris where the measures are taken.

airparif$datasup describes the 101 days with meteorological variables.
}
\examples{
  data(airparif)
}
\keyword{datasets}

