% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{BuildSTCdata}
\alias{BuildSTCdata}
\title{Constructors of the class \linkS4class{STCdata}}
\usage{
BuildSTCdata(x, map, m = 1:(dim(x)[3]))
}
\arguments{
\item{x}{array It contains the observations to cluster where the dimesions are respectively: number of the observation, site of the observation, time of the observation.}

\item{map}{matrix. It gives the spatial coordiantes of each site.}

\item{m}{numeric. It indicates the moments of observations.}
}
\value{
Returns an instance of \linkS4class{STCdata}.
}
\description{
Constructors of the class \linkS4class{STCdata}
}

