% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,STCresults-method}
\title{plot function.}
\usage{
\S4method{plot}{STCresults}(x, y, ...)
}
\arguments{
\item{object}{instance of \code{\linkS4class{STCresults}}.}
}
\description{
This function gives the plot of an instance of \code{\linkS4class{STCresults}}.
}

