% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,STCresults-method}
\title{Summary function.}
\usage{
\S4method{print}{STCresults}(x)
}
\arguments{
\item{x}{an instance of \code{\linkS4class{STCresults}}.}
}
\description{
This function prints the elements  of an instance of \code{\linkS4class{STCresults}}..
}

