% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTimeClus.R
\name{spatimeclass}
\alias{spatimeclass}
\title{This function performs classifies a sample}
\usage{
spatimeclass(ech, model, param)
}
\arguments{
\item{ech}{\linkS4class{STCdata} It contains the observations to classify (mandatory).}

\item{model}{\linkS4class{STCmodel}. It defines the model at hand (mandatory).}

\item{param}{list of \linkS4class{STCparam}. It gives the model parameters (mandatory).}
}
\value{
Returns an instance of \linkS4class{STCresults}.
}
\description{
This function performs classifies a sample
}

