% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{STCcriteria-class}
\alias{STCcriteria-class}
\title{Constructor of [\code{\linkS4class{STCcriteria}}] class}
\description{
\describe{
  \item{loglike}{numeric. It indicates the value of the log-likelihood.}
  \item{AIC}{numeric. It indicates the value of the AIC criterion.}
  \item{BIC}{numeric. It indicates the value of the BIC criterion.}
  \item{ICL}{numeric. It indicates the value of the ICL criterion.}
}
}
\examples{
  getSlots("STCcriteria")

}

