% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{STCdata-class}
\alias{STCdata-class}
\title{Constructor of [\code{\linkS4class{STCdata}}] class}
\description{
\describe{
  \item{x}{matrix. It contains the observations. Each column corresponds to an individual. The row indicates the values of each site for each time.}
  \item{TT}{numeric. It indicates the number of elements of the time grid.}
  \item{JJ}{numeric. It indicates the number of sites.}
  \item{n}{numeric. It indicates the number of observations.}
  \item{map}{numeric. It indicates the spatial coordinates of each site.}
}
}
\examples{
  getSlots("STCdata")

}

