% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{STCmodel-class}
\alias{STCmodel-class}
\title{Constructor of [\code{\linkS4class{STCmodel}}] class}
\description{
\describe{
  \item{G}{integer. It defines the number of mixture components.}
  \item{K}{integer. It defines the number of polynoms for each component.}
  \item{Q}{integer. It defines the degree of the polynoms.}
  \item{spatial}{integer.  It defines if the spatial dependencies are considered (1:yes, 0:no).}
  \item{nbparam}{integer.  It indicates the number of parameters involved by the model.}
}
}
\examples{
  getSlots("STCmodel")

}

