% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{STCmodel}
\alias{STCmodel}
\title{This function builts an instance of \linkS4class{STCmodel}.}
\usage{
STCmodel(G, K, Q, nospatial)
}
\arguments{
\item{G}{integer. It defines the number of mixture components.}

\item{K}{integer. It defines the number of polynoms for each component.}

\item{Q}{integer. It defines the degree of the polynoms.}

\item{nospatial}{binary. It defines if the spatial dependencies are considered (1:no, 0:yes).}
}
\value{
Returns an instance of \linkS4class{STCmodel}.
}
\description{
This function builts an instance of \linkS4class{STCmodel}.
}

