% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{STCparam-class}
\alias{STCparam-class}
\title{Constructor of [\code{\linkS4class{STCparam}}] class}
\description{
\describe{
  \item{proportions}{numeric. It defines the component proportions.}
  \item{lambda}{list. It defines the logistic coefficients per component.}
  \item{beta}{list.  It defines the polynom coefficients per component.}
  \item{sigma}{matrix.  It defines the variance associated to each polynom per component.}
}
}
\examples{
  getSlots("STCparam")

}

