% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{STCpartitions-class}
\alias{STCpartitions-class}
\title{Constructor of [\code{\linkS4class{STCpartitions}}] class}
\description{
\describe{
  \item{hardind}{numeric. It indicates the hard partition of the individuals (obtained by the MAP rule qpplied with the MLE).}
  \item{fuzzyind}{matrix. It indicates the fuzzy partition (conditional probability of the component membership) of the individuals.}
  \item{hardseg}{list. It indicates the segmentation (most probable polynom according to the spatial and temporal grid) per components}
}
}
\examples{
  getSlots("STCpartitions")

}

