% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{STCresults-class}
\alias{STCresults-class}
\title{Constructor of [\code{\linkS4class{STCresults}}] class}
\description{
\describe{
  \item{model}{\linkS4class{STCmodel}. It contains the elements relied to the model.}
  \item{data}{\linkS4class{STCdata}. It contains the elements relied to the data.}
  \item{param}{\linkS4class{STCparam}. It contains the elements relied to the parameters.}
  \item{criteria}{\linkS4class{STCcriteria}. It contains the elements relied to the information criteria.}
  \item{partitions}{\linkS4class{STCpartitions}. It contains the elements relied to the partitions.}
  \item{tune}{\linkS4class{STCtune}. It contains the tunning parameters of the algorithm.}
  \item{allmodels}{matrix. list of the estimnated models and their information criterion.}
}
}
\examples{
  getSlots("STCresults")

}

