% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{STCtune-class}
\alias{STCtune-class}
\title{Constructor of [\code{\linkS4class{STCtune}}] class}
\description{
\describe{
  \item{tol}{numeric. The algorithm is stopped when two successive iterations increase the log-likelihood less than tol.}
  \item{nbinitSmall}{numeric. Number of random initializations for the short run EM algorithm.}
  \item{nbinitKept}{numeric. Number of initializations kept for the long run EM algorithm.}
  \item{nbiterSmall}{numeric. Maximum number of iteration before stopping the short run EM algorithm.}
  \item{nbiterKept}{numeric. Maximum number of iteration before stopping the long run EM algorithm.}
}
}
\examples{
  getSlots("STCtune")

}

