% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,STCresults-method}
\title{Summary function.}
\usage{
\S4method{summary}{STCresults}(object)
}
\arguments{
\item{object}{instance of \code{\linkS4class{STCresults}}.}
}
\description{
This function gives the summary of an instance of \code{\linkS4class{STCresults}}.
}

