% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTimeClus.R
\docType{package}
\name{SpaTimeClus-package}
\alias{SpaTimeClus}
\alias{SpaTimeClus-package}
\title{SpaTimeClus a package for clustering spatio-temporal data}
\description{
todo
}
\details{
\tabular{ll}{
  Package: \tab SpaTimeClus\cr 
  Type: \tab Package\cr 
  Version: \tab 1.0.0\cr
  Date: \tab 2015-11-03\cr 
  License: \tab GPL-2\cr 
  LazyLoad: \tab yes\cr
}

todo
}
\examples{
data(ech)

output <- spatimeclus(ech = ech, G = 2, K = 2, Q = 2, crit = "BIC", nbcores = 6)

summary(output)

print(output)

plot(output) 

}
\author{
Author: Marbac M., and McNicholas P.
}
\references{
Marbac M., and McNicholas P.
}
\keyword{package}

