% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTimeClus.R
\name{spatimeclus}
\alias{spatimeclus}
\title{This function performs the maximum likelihood estimation for a known model in clustering}
\usage{
spatimeclus(ech, G, K, Q, crit = "BIC", tol = 0.1, param = NULL,
  nbcores = 1, nbinitSmall = 100, nbinitKept = 10, nbiterSmall = 10,
  nbiterKept = 500)
}
\arguments{
\item{ech}{\linkS4class{STCdata} It contains the observations to cluster (mandatory).}

\item{G}{numeric. It defines possible numbers of components.}

\item{K}{numeric. It defines possible numbers of regressions per components}

\item{Q}{numeric. It defines possible degrees of regressions.}

\item{crit}{character. It indicates the criterion used for the model selection ("AIC", "BIC" or "ICL", optional, default is "BIC").}

\item{tol}{numeric. The algorithm is stopped when the loglikelihood increases less than tol during two successive iterations (optional, default is 0.1).}

\item{param}{list of \linkS4class{STCparam}. It gives the initial values of the EM algorithm (optional).}

\item{nbcores}{numeric.  It defines the numerber of cores used by the alogrithm, only for Linux and Mac (optional, default is 1).}

\item{nbinitSmall}{numeric. It defines the number of random initializations (optional, default is 100).}

\item{nbinitKept}{numeric. It defines the number of chains estimated until convergence (optional, default is 10).}

\item{nbiterSmall}{numeric. It defines the number of iterations before keeping the nbinitKept best chains (optional, default is 10).}

\item{nbiterKept}{numeric. It defines the maximum number of iterations before to stop the algorith; (optional, default is 500).}
}
\value{
Returns an instance of \linkS4class{STCresults}.
}
\description{
This function performs the maximum likelihood estimation for a known model in clustering
}
\examples{
data(ech)

output <- spatimeclus(ech = ech, G = 2, K = 2, Q = 2, crit = "BIC", nbcores = 6)

summary(output)

print(output)

plot(output)


}

